Move ordered_dict.py outside of simplejson
